requirements.txt
kornia==0.6.12
librosa==0.10.0.post2
matplotlib==3.3.1
nltk==3.8.1
numpy==1.23.0
omegaconf==2.3.0
pandas==2.0.0
psutil==5.9.4
psutil==5.9.0
pydub==0.25.1
pytorch_lightning==1.8.6
rich==13.7.1
scipy==1.9.1
soundfile==0.12.1
TextGrid==1.5
torch==2.1.2
torchmetrics==0.7.0
tqdm==4.65.0
transformers==4.27.4
openai-whisper
